UPDATE / friends routes
